Break City Filters into component
